+++ packages/cli/lib/commands/info.js
@@ -1,13 +1,12 @@
-const envinfo = require('envinfo')
+const envinfo = require("envinfo");
 
 module.exports = async () => {
-  const data = await envinfo
-    .run({
-      System: ['OS', 'CPU'],
-      Binaries: ['Node', 'Yarn', 'npm'],
-      Browsers: ['Chrome', 'Edge', 'Firefox', 'Safari'],
-      npmPackages: '?(@)gridsome{-*,/*,}',
-      npmGlobalPackages: ['@gridsome/cli']
-    })
-  console.log(data)
-}
+  const data = await envinfo.run({
+    System: ["OS", "CPU"],
+    Binaries: ["Node", "Yarn", "npm"],
+    Browsers: ["Chrome", "Edge", "Firefox", "Safari"],
+    npmPackages: "?(@)gridsome{-*,/*,}",
+    npmGlobalPackages: ["@zuckersalzundpfeffer/gridsome-cli"],
+  });
+  console.log(data);
+};